Title: 실수 받아 그대로 출력, Time: 45ms, Memory: 16MB, Status: Passed - Codetree

+++ 250422/실수 받아 그대로 출력/enter-real-value-and-print.py
@@ -1,2 +1,2 @@
-n = int(input())
+n = float(input())
 print(f'{n:.2f}')
